brolly/requirements.txt
asn1crypto==0.24.0
blinker==1.3
boto==2.48.0
boto3==1.5.24
botocore==1.8.38
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
Click==7.0
configobj==5.0.6
confusable-homoglyphs==3.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==1.2.3
Cython==0.27.3
Django==2.0.2
django-cors-headers==2.1.0
django-filter==1.1.0
django-jet==1.0.7
django-registration==2.4.1
django-rest-swagger==2.1.2
django-storages==1.6.5
django-summernote==0.8.8.5
django-super-inlines==0.1.4
djangorestframework==3.7.7
docutils==0.14
feedparser==5.2.1
Flask==1.0.2
Flask-MySQLdb==0.2.0
google-api-python-client==1.4.1
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonpatch==1.10
jsonpointer==1.9
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.0
oauth2client==4.1.2
oauthlib==1.0.3
openapi-codec==1.3.2
opencv-python==3.4.0.12
pbr==3.1.1
Pillow==5.0.0
prettytable==0.7.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycurl==7.43.0
PyJWT==1.3.0
pyserial==3.0.1
pystan==2.17.1.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
simplejson==3.13.2
six==1.11.0
stevedore==1.28.0
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
Werkzeug==0.14.1
